requirements.txt
torch>=2.0.0
transformers>=4.36.0
streamlit>=1.30.0
Pillow
opencv-python
numpy
pyttsx3
ultralytics
